:recycle: [STMT-184] 활동 조회에 대한 로직을 Query / UseCase로 분리
