+++ src/test/java/com/stumeet/server/activity/adapter/in/ActivityQueryApiTest.java
@@ -63,10 +63,10 @@ class ActivityQueryApiTest extends ApiTest {
                                     fieldWithPath("data.content").description("활동 내용"),
                                     fieldWithPath("data.imageUrl[].id").description("활동 이미지의 아이디"),
                                     fieldWithPath("data.imageUrl[].imageUrl").description("활동 이미지의 URL"),
-                                    fieldWithPath("data.author.memberId").description("활동 작성자의 멤버 ID"),
+                                    fieldWithPath("data.author.memberId").description("활동 작성자 ID"),
                                     fieldWithPath("data.author.name").description("활동 작성자 이름"),
                                     fieldWithPath("data.author.profileImageUrl").description("활동 작성자 프로필 이미지 URL"),
-                                    fieldWithPath("data.participants[].memberId").description("참여자의 멤버 ID"),
+                                    fieldWithPath("data.participants[].memberId").description("참여자 ID"),
                                     fieldWithPath("data.participants[].name").description("참여자 이름"),
                                     fieldWithPath("data.participants[].profileImageUrl").description("참여자 프로필 이미지 URL"),
                                     fieldWithPath("data.status").description("나의 활동 상태"),
